--HG--
branch : trunk

+++ fileprocess/fileprocess/actions/cleanup.py
@@ -1,7 +1,7 @@
 import logging
 from baseaction import BaseAction
 import os
-import fileprocess
+from ..processingthread import UploadStatus, na
 
 log = logging.getLogger(__name__)
 
@@ -12,13 +12,13 @@ class Cleanup(BaseAction):
     terminal action.
     """
     def process(self, file):
-        fileprocess.UploadStatus(
+        us = UploadStatus(
             file.get('msg', "File Successfully Uploaded"),
-            file.get('na', fileprocess.na.NOTHING),
+            file.get('na', na.NOTHING),
             file
         )
+        log.debug('Cleanup %s:%s', file.get('msg'), file.get('na'))
 
-        try:
+        if os.path.exists(file.get('fname')):
+            log.debug('Removing %s:%s', file.get('title'), file.get('fname'))
             os.remove(file['fname'])
-        except:
-            pass
